save ALL of the autocheck edits

+++ SQLServerDB/mapping_note_Table.cs
@@ -213,6 +213,9 @@ namespace SQLServerDB
         /// </summary>
         public void Clear_Database_Table()
         {
+            if (CountRows() == 0)
+                return;
+
             string strCommand = "DELETE FROM " + theTable;
             if (!DBUtils.ExecuteSqlNonQuery(strCommand))
                 LogManager.writeToLog(" ExecuteSqlNonQuery returned: FALSE; in mapping_note_Table.cs:Clear_Database_Table.ExecuteSqlNonQuery()");
