+++ SQLServerDB/team_note_Table.cs
@@ -245,6 +245,9 @@ namespace SQLServerDB
         /// </summary>
         public void Clear_Database_Table()
         {
+            if (CountRows() == 0)
+                return;
+
             string strCommand = "DELETE FROM " + theTable;
             if (!DBUtils.ExecuteSqlNonQuery(strCommand))
                 LogManager.writeToLog(" ExecuteSqlNonQuery returned: FALSE; in team_note_Table.cs:Clear_Database_Table.ExecuteSqlNonQuery()");
@@ -306,6 +309,7 @@ namespace SQLServerDB
         /// <returns></returns>
         public int CountRows()
         {
+
             SqlConnection myConnection = DBUtils.GetNewSqlConnection();
             if (myConnection == null)
             {
